// src/components/common/Tours/tours/createAction/enums.ts
export enum CreateActionTourTargets {
  CreatingActionsPanel = 'createActionCreatingActions',
  EnterTitleField = 'createActionEnterTitle',
  ChooseActionButton = 'createActionChooseAction',
  SelectActionButton = 'createActionSelectAction',
  CreateSimplePaymentButton = 'createActionCreateSimplePayment',
  ActionOverviewSection = 'createActionActionOverview',
  CompleteDetailsSection = 'createActionCompleteDetails',
  DecisionMethodField = 'createActionDecisionMethod',
  ActionDescriptionField = 'createActionActionDescription',
  CancelActionButton = 'createActionCancelAction',
  CreateActionButton = 'createActionCreateAction',
}